application_code/requirements.txt
aiohttp==3.12.14
gradio==5.35.0
langfuse==3.1.3
nest-asyncio==1.6.0
openai==1.93.1
openai-agents==0.1.0
pydantic==2.11.7
pydantic-ai-slim[logfire]==0.3.7
weaviate-client==4.15.4
python-dotenv==1.1.1
langchain-mcp-adapters==0.1.9
atlassian-python-api==4.0.4
html2text==2025.4.15
beautifulsoup4==4.13.4
python-dotenv==1.1.1
langchain==0.3.27
langgraph==0.5.4
langchain-openai==0.3.28
atlassian-python-api==4.0.4
mcp==1.11.0
html2text==2025.4.15